feat: factor out common.ts and errors.ts into a `utils` package
* There's a bunch of TypeScript utilities that we can use across the
  monorepo here
