+++ apps/frontend/package.json
@@ -27,6 +27,7 @@
     "@mui/icons-material": "^5.15.6",
     "@mui/material": "^5.15.6",
     "@mui/x-date-pickers": "^6.19.0",
+    "@opensource-observer/utils": "workspace:*",
     "@plasmicapp/loader-nextjs": "^1.0.360",
     "@segment/analytics-next": "^1.66.0",
     "@supabase/auth-helpers-nextjs": "^0.8.7",
