+++ lib/utils/.eslintrc.json
@@ -1,0 +1,7 @@
+{
+  "extends": ["../../.eslintrc.js"],
+  "root": false,
+  "parserOptions": {
+    "project": ["./lib/utils/tsconfig.json"]
+  }
+}
