+++ lib/utils/src/common.ts
@@ -1,4 +1,4 @@
-import { NullOrUndefinedValueError, AssertionError } from "./types/errors";
+import { NullOrUndefinedValueError, AssertionError } from "./errors.js";
 /**
  * Explicitly marks a promise as something we won't await
  * @param _promise
