+++ ops/external-prs/package.json
@@ -36,6 +36,7 @@
   "dependencies": {
     "@google-cloud/bigquery": "^7.5.1",
     "@opensource-observer/oss-artifact-validators": "workspace:*",
+    "@opensource-observer/utils": "workspace:*",
     "@types/columnify": "^1.5.4",
     "@types/libsodium-wrappers": "^0.7.13",
     "@types/mustache": "^4.2.5",
