+++ pnpm-workspace.yaml
@@ -5,5 +5,6 @@ packages:
   - "apps/frontend"
   - "apps/hasura"
   - "lib/oss-artifact-validators"
+  - "lib/utils"
   - "ops/external-prs"
   - "ops/github-app-utils"
